feat: extract github-specific repo auto-import functionality
