+++ internal/jobs/repo/import.go
@@ -2,22 +2,15 @@ package repo
 
 import (
 	"context"
-	"database/sql"
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"time"
 
-	"github.com/google/go-github/v41/github"
-	"github.com/google/uuid"
-	"github.com/jackc/pgtype"
 	"github.com/jackc/pgx/v4"
 	"github.com/jackc/pgx/v4/pgxpool"
 	"github.com/mergestat/mergestat/internal/db"
 	"github.com/mergestat/sqlq"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
-	"golang.org/x/oauth2"
 )
 
 type githubRepository struct {
@@ -30,14 +23,13 @@ func (repo *githubRepository) URL() string {
 	return fmt.Sprintf("https://github.com/%s/%s", repo.Owner, repo.Name)
 }
 
-type fetchFunc func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error)
-
 // AutoImport implements the githubRepository auto-import job to automatically
 // sync githubRepository from user- or org- accounts.
-func AutoImport(l *zerolog.Logger, pool *pgxpool.Pool) sqlq.HandlerFunc {
+func AutoImport(pool *pgxpool.Pool) sqlq.HandlerFunc {
 	var queries = db.New(pool)
 
 	return func(ctx context.Context, job *sqlq.Job) (err error) {
+		var l = zerolog.Ctx(ctx)
 		var logger = job.Logger()
 
 		// start sending periodic keep-alive pings!
@@ -66,7 +58,13 @@ func AutoImport(l *zerolog.Logger, pool *pgxpool.Pool) sqlq.HandlerFunc {
 			// if the execution fails for some reason, only that import is marked as failed
 			// the job still continues executing.
 			var importError error
-			if importError = handleImport(ctx, queries.WithTx(tx), imp); importError != nil {
+			if imp.VendorName == "github" {
+				importError = handleGithubImport(ctx, queries.WithTx(tx), imp)
+			} else {
+				importError = errors.Errorf("unknown vendor: %s", imp.VendorName)
+			}
+
+			if importError != nil {
 				logger.Warnf("import(%s) failed: %v", imp.ID, importError.Error())
 				l.Warn().Msgf("import(%s) failed: %v", imp.ID, importError.Error())
 				_ = tx.Rollback(ctx)
@@ -96,117 +94,6 @@ func AutoImport(l *zerolog.Logger, pool *pgxpool.Pool) sqlq.HandlerFunc {
 	}
 }
 
-// handleImport handles execution of a given import configuration
-func handleImport(ctx context.Context, qry *db.Queries, imp db.ListRepoImportsDueForImportRow) (err error) {
-	var token string
-	if token, err = qry.FetchCredential(ctx, imp.Provider); err != nil {
-		return err
-	}
-
-	var public = token == ""
-
-	var client *github.Client
-	if !public {
-		var tokenSource = oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})
-		client = github.NewClient(oauth2.NewClient(ctx, tokenSource))
-	} else {
-		client = github.NewClient(&http.Client{})
-	}
-
-	var settings struct {
-		Type               string   `json:"type"`
-		Login              string   `json:"userOrOrg"`
-		RemoveDeletedRepos bool     `json:"removeDeletedRepos"`
-		DefaultSyncTypes   []string `json:"defaultSyncTypes"`
-	}
-
-	if err = json.Unmarshal(imp.Settings.Bytes, &settings); err != nil {
-		return errors.Wrapf(err, "failed to parse import settings")
-	}
-
-	var fetchFunction fetchFunc
-	switch settings.Type {
-	case "GITHUB_ORG":
-		var opts = &github.RepositoryListByOrgOptions{}
-		if public {
-			opts.Type = "public"
-		}
-		fetchFunction = fetchByOrg(client, settings.Login, opts)
-	case "GITHUB_USER":
-		var opts = &github.RepositoryListOptions{}
-		if public {
-			opts.Type = "public"
-		}
-		fetchFunction = fetchByUser(client, settings.Login, opts)
-	default:
-		return errors.Errorf("unknown import type: %s", settings.Type)
-	}
-
-	var repos []*githubRepository
-	if repos, err = fetchRepositories(ctx, fetchFunction); err != nil {
-		return errors.Wrapf(err, "failed to fetch repositories")
-	}
-
-	// remove any deleted repositories
-	if settings.RemoveDeletedRepos {
-		var params = db.DeleteRemovedReposParams{Column1: imp.ID, Column2: repoUrls(repos, settings.Login)}
-		if err = qry.DeleteRemovedRepos(ctx, params); err != nil {
-			return errors.Wrapf(err, "failed to remove deleted repositories")
-		}
-	}
-
-	// capture list of existing repositories before we do the upsert below
-	var existing []string
-	if len(settings.DefaultSyncTypes) > 0 { // only if defaultSyncTypes is configured
-		if existing, err = qry.GetRepoUrlFromImport(ctx, imp.ID); err != nil {
-			return errors.Wrapf(err, "failed to enable default sync")
-		}
-	}
-
-	// upsert all fetched repositories
-	for _, repo := range repos {
-		var opts = db.UpsertRepoParams{
-			Repo:         fmt.Sprintf("https://github.com/%s/%s", settings.Login, repo.Name),
-			IsGithub:     sql.NullBool{Bool: true, Valid: true},
-			RepoImportID: uuid.NullUUID{Valid: true, UUID: imp.ID},
-			Tags:         pgtype.JSONB{Status: pgtype.Present, Bytes: []byte(repo.Topics)},
-		}
-
-		if err = qry.UpsertRepo(ctx, opts); err != nil {
-			return errors.Wrapf(err, "failed to upsert repository")
-		}
-	}
-
-	// (optional) configure default sync types
-	if len(settings.DefaultSyncTypes) > 0 {
-		// batch is a collection of newly added repositories
-		var batch = difference(existing, repoUrls(repos, settings.Login))
-
-		// convert batch into a collection of repo ids
-		var ids []uuid.UUID
-		if ids, err = qry.GetRepoIDsFromRepoImport(ctx, db.GetRepoIDsFromRepoImportParams{Importid: imp.ID, Reposurls: batch}); err != nil {
-			return errors.Wrapf(err, "failed to enable default sync")
-		}
-
-		// for each new repo, enable the provided syncs
-		for _, id := range ids {
-			for _, syncType := range settings.DefaultSyncTypes {
-				var params = db.InsertNewDefaultSyncParams{Repoid: id, Synctype: syncType}
-				if err = qry.InsertNewDefaultSync(ctx, params); err != nil {
-					return errors.Wrapf(err, "failed to enable default sync")
-				}
-			}
-		}
-
-		// enqueue all newly added syncs
-		if err = qry.EnqueueAllSyncs(ctx); err != nil {
-			return errors.Wrapf(err, "failed to enable default sync")
-		}
-	}
-
-	return qry.MarkRepoImportAsUpdated(ctx, imp.ID)
-}
-
 func repoUrls(repos []*githubRepository, repoOwner string) []string {
 	var ret = make([]string, len(repos))
 	for i, repo := range repos {
@@ -234,47 +121,4 @@ func difference(existing, new []string) []string {
 	}
 
 	return diff
-}
-
-func fetchRepositories(ctx context.Context, fetch fetchFunc) (_ []*githubRepository, err error) {
-	var repositories []*githubRepository
-	var repos []*github.Repository
-	var resp *github.Response
-
-	var page = 1
-	for {
-		if repos, resp, err = fetch(ctx, page); err != nil {
-			return repositories, err
-		}
-
-		for _, repo := range repos {
-			var jsonStr []byte
-			if jsonStr, err = json.Marshal(&repo.Topics); err != nil {
-				return repositories, err
-			}
-
-			repositories = append(repositories, &githubRepository{Name: *repo.Name, Owner: *repo.Owner.OrganizationsURL,
-				Topics: string(jsonStr)})
-		}
-
-		if resp.NextPage <= page {
-			break
-		}
-		page = resp.NextPage
-	}
-	return repositories, err
-}
-
-func fetchByUser(client *github.Client, user string, opts *github.RepositoryListOptions) fetchFunc {
-	return func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error) {
-		opts.Page = page
-		return client.Repositories.List(ctx, user, opts)
-	}
-}
-
-func fetchByOrg(client *github.Client, org string, opts *github.RepositoryListByOrgOptions) fetchFunc {
-	return func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error) {
-		opts.Page = page
-		return client.Repositories.ListByOrg(ctx, org, opts)
-	}
 }
