+++ internal/jobs/repo/vendor_github.go
@@ -1,0 +1,170 @@
+package repo
+
+import (
+	"context"
+	"database/sql"
+	"encoding/json"
+	"fmt"
+	"github.com/google/go-github/v41/github"
+	"github.com/google/uuid"
+	"github.com/jackc/pgtype"
+	"github.com/mergestat/mergestat/internal/db"
+	"github.com/pkg/errors"
+	"golang.org/x/oauth2"
+	"net/http"
+)
+
+type fetchFunc func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error)
+
+func handleGithubImport(ctx context.Context, qry *db.Queries, imp db.ListRepoImportsDueForImportRow) (err error) {
+	var token string
+	if token, err = qry.FetchCredential(ctx, imp.Provider); err != nil {
+		return err
+	}
+
+	var public = token == ""
+
+	var client *github.Client
+	if !public {
+		var tokenSource = oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})
+		client = github.NewClient(oauth2.NewClient(ctx, tokenSource))
+	} else {
+		client = github.NewClient(&http.Client{})
+	}
+
+	var settings struct {
+		Type               string   `json:"type"`
+		Login              string   `json:"userOrOrg"`
+		RemoveDeletedRepos bool     `json:"removeDeletedRepos"`
+		DefaultSyncTypes   []string `json:"defaultSyncTypes"`
+	}
+
+	if err = json.Unmarshal(imp.Settings.Bytes, &settings); err != nil {
+		return errors.Wrapf(err, "failed to parse import settings")
+	}
+
+	var fetchFunction fetchFunc
+	switch settings.Type {
+	case "GITHUB_ORG":
+		var opts = &github.RepositoryListByOrgOptions{}
+		if public {
+			opts.Type = "public"
+		}
+		fetchFunction = fetchByOrg(client, settings.Login, opts)
+	case "GITHUB_USER":
+		var opts = &github.RepositoryListOptions{}
+		if public {
+			opts.Type = "public"
+		}
+		fetchFunction = fetchByUser(client, settings.Login, opts)
+	default:
+		return errors.Errorf("unknown import type: %s", settings.Type)
+	}
+
+	var repos []*githubRepository
+	if repos, err = fetchRepositories(ctx, fetchFunction); err != nil {
+		return errors.Wrapf(err, "failed to fetch repositories")
+	}
+
+	// remove any deleted repositories
+	if settings.RemoveDeletedRepos {
+		var params = db.DeleteRemovedReposParams{Column1: imp.ID, Column2: repoUrls(repos, settings.Login)}
+		if err = qry.DeleteRemovedRepos(ctx, params); err != nil {
+			return errors.Wrapf(err, "failed to remove deleted repositories")
+		}
+	}
+
+	// capture list of existing repositories before we do the upsert below
+	var existing []string
+	if len(settings.DefaultSyncTypes) > 0 { // only if defaultSyncTypes is configured
+		if existing, err = qry.GetRepoUrlFromImport(ctx, imp.ID); err != nil {
+			return errors.Wrapf(err, "failed to enable default sync")
+		}
+	}
+
+	// upsert all fetched repositories
+	for _, repo := range repos {
+		var opts = db.UpsertRepoParams{
+			Repo:         fmt.Sprintf("https://github.com/%s/%s", settings.Login, repo.Name),
+			IsGithub:     sql.NullBool{Bool: true, Valid: true},
+			RepoImportID: uuid.NullUUID{Valid: true, UUID: imp.ID},
+			Tags:         pgtype.JSONB{Status: pgtype.Present, Bytes: []byte(repo.Topics)},
+		}
+
+		if err = qry.UpsertRepo(ctx, opts); err != nil {
+			return errors.Wrapf(err, "failed to upsert repository")
+		}
+	}
+
+	// (optional) configure default sync types
+	if len(settings.DefaultSyncTypes) > 0 {
+		// batch is a collection of newly added repositories
+		var batch = difference(existing, repoUrls(repos, settings.Login))
+
+		// convert batch into a collection of repo ids
+		var ids []uuid.UUID
+		if ids, err = qry.GetRepoIDsFromRepoImport(ctx, db.GetRepoIDsFromRepoImportParams{Importid: imp.ID, Reposurls: batch}); err != nil {
+			return errors.Wrapf(err, "failed to enable default sync")
+		}
+
+		// for each new repo, enable the provided syncs
+		for _, id := range ids {
+			for _, syncType := range settings.DefaultSyncTypes {
+				var params = db.InsertNewDefaultSyncParams{Repoid: id, Synctype: syncType}
+				if err = qry.InsertNewDefaultSync(ctx, params); err != nil {
+					return errors.Wrapf(err, "failed to enable default sync")
+				}
+			}
+		}
+
+		// enqueue all newly added syncs
+		if err = qry.EnqueueAllSyncs(ctx); err != nil {
+			return errors.Wrapf(err, "failed to enable default sync")
+		}
+	}
+
+	return qry.MarkRepoImportAsUpdated(ctx, imp.ID)
+}
+
+func fetchRepositories(ctx context.Context, fetch fetchFunc) (_ []*githubRepository, err error) {
+	var repositories []*githubRepository
+	var repos []*github.Repository
+	var resp *github.Response
+
+	var page = 1
+	for {
+		if repos, resp, err = fetch(ctx, page); err != nil {
+			return repositories, err
+		}
+
+		for _, repo := range repos {
+			var jsonStr []byte
+			if jsonStr, err = json.Marshal(&repo.Topics); err != nil {
+				return repositories, err
+			}
+
+			repositories = append(repositories, &githubRepository{Name: *repo.Name, Owner: *repo.Owner.OrganizationsURL,
+				Topics: string(jsonStr)})
+		}
+
+		if resp.NextPage <= page {
+			break
+		}
+		page = resp.NextPage
+	}
+	return repositories, err
+}
+
+func fetchByUser(client *github.Client, user string, opts *github.RepositoryListOptions) fetchFunc {
+	return func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error) {
+		opts.Page = page
+		return client.Repositories.List(ctx, user, opts)
+	}
+}
+
+func fetchByOrg(client *github.Client, org string, opts *github.RepositoryListByOrgOptions) fetchFunc {
+	return func(ctx context.Context, page int) ([]*github.Repository, *github.Response, error) {
+		opts.Page = page
+		return client.Repositories.ListByOrg(ctx, org, opts)
+	}
+}
